Move caching code to hetmat.caching module
